requirements.txt
streamlit>=1.36.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
typer>=0.9.0
rich>=13.0.0
tqdm>=4.65.0
requests>=2.31.0
httpx>=0.27.0
matplotlib>=3.8.0
